+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [UIRS-30](https://issues.folio.org/browse/UIRS-30) Compile Translation Files into AST Format.
 * [UIRS-41](https://issues.folio.org/browse/UIRS-41) Add Accession holdings workflow preference
 * [UIRS-42](https://issues.folio.org/browse/UIRS-42) Update "remote-storage-mappings" interface
+* [UIRS-38](https://issues.folio.org/browse/UIRS-38) Update remote storage permissions
 
 ## [1.0.0](https://github.com/folio-org/ui-remote-storage/tree/v1.0.0) (2021-03-19)