+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change history for ui-remote-storage
 
 ## (IN PROGRESS)
+
+## [1.1.0](https://github.com/folio-org/ui-remote-storage/tree/v1.1.0) (2021-06-17)
+[Full Changelog](https://github.com/folio-org/ui-remote-storage/compare/v1.0.0...v1.1.0)
+
 * [UIRS-15](https://issues.folio.org/browse/UIRS-15) Remote Storage Settings: Credential properties (apiKey)
 * [UIRS-24](https://issues.folio.org/browse/UIRS-24) Add Returning workflow preference
 * [UIRS-27](https://issues.folio.org/browse/UIRS-27) Add to "Remote storage: Create, read, update, delete" all the missing permissions on configurations and providers